Update crea worker

+++ src/app/infrastructure/background/worker-heavy-task.adapter.ts
@@ -1,21 +1,31 @@
 import { Injectable } from '@angular/core';
 import { HeavyTaskPort } from 'src/app/core/ports/heavy-task.port';
 
-
 @Injectable()
 export class WorkerHeavyTaskAdapter implements HeavyTaskPort {
 
-  calculateSqrt(n: number): Promise<number> {
+  calculateSqrt(
+    n: number,
+    onProgress?: (value: number) => void
+  ): Promise<number> {
 
     return new Promise((resolve) => {
 
       const worker = new Worker(
-        new URL('./sqrt.worker', import.meta.url)
+        new URL('./sqrt.worker', import.meta.url),
+        { type: 'module' }
       );
 
       worker.onmessage = ({ data }) => {
-        resolve(data);
-        worker.terminate();
+
+        if (data.progress !== undefined) {
+          onProgress?.(data.progress);
+        }
+
+        if (data.done) {
+          resolve(data.result);
+          worker.terminate();
+        }
       };
 
       worker.postMessage(n);
